requirements.txt
fastapi
uvicorn
pytest
sqlalchemy
pydentic
passlib[bcrypt]
python-jose
httpx
python-dotenv
python-multipart